fix(web): correct status colors, date-range close, and move RunStatusBadge
- Only auto-close DatePickerRange popover when a distinct end day is
  picked; react-day-picker returns {from, to} equal to the same day on
  the first click, which previously closed the popover before a real
  range could be selected.
- Give --status-* tokens explicit theme-appropriate hues instead of
  aliasing the non-semantic --chart-* tokens, which produced wrong
  colors per theme (e.g. yellow/purple ERROR).
- Point WARN/ERROR/failed/bad-delta at semantic destructive/idle
  tokens instead of chart tokens that don't map to amber/red.
- Move RunStatusBadge into web/components/ds/ (renamed from the old
  StatusBadge in task-run-analytics-view.tsx, which Phase 6 deletes)
  and update its call sites.
