+++ web/app/globals.css
@@ -36,12 +36,12 @@
   --chart-3: oklch(0.398 0.07 227.392);
   --chart-4: oklch(0.828 0.189 84.429);
   --chart-5: oklch(0.769 0.188 70.08);
-  --status-connected: var(--chart-2);
-  --status-streaming: var(--chart-1);
-  --status-provisioning: var(--chart-5);
-  --status-idle: var(--chart-3);
+  --status-connected: oklch(0.6 0.118 184.704);
+  --status-streaming: oklch(0.6 0.118 184.704);
+  --status-provisioning: oklch(0.546 0.215 262.881);
+  --status-idle: oklch(0.769 0.188 70.08);
   --status-offline: var(--muted-foreground);
-  --status-error: var(--chart-4);
+  --status-error: var(--destructive);
   --heatmap-1: oklch(0.86 0.08 250);
   --heatmap-2: oklch(0.72 0.13 250);
   --heatmap-3: oklch(0.59 0.17 250);
@@ -83,12 +83,12 @@
   --chart-3: oklch(0.769 0.188 70.08);
   --chart-4: oklch(0.627 0.265 303.9);
   --chart-5: oklch(0.645 0.246 16.439);
-  --status-connected: var(--chart-2);
-  --status-streaming: var(--chart-1);
-  --status-provisioning: var(--chart-5);
-  --status-idle: var(--chart-3);
+  --status-connected: oklch(0.696 0.17 162.48);
+  --status-streaming: oklch(0.696 0.17 162.48);
+  --status-provisioning: oklch(0.623 0.214 259.815);
+  --status-idle: oklch(0.769 0.188 70.08);
   --status-offline: var(--muted-foreground);
-  --status-error: var(--chart-4);
+  --status-error: var(--destructive);
   --heatmap-1: oklch(0.35 0.09 250);
   --heatmap-2: oklch(0.45 0.13 250);
   --heatmap-3: oklch(0.56 0.16 250);
