+++ web/components/ds/index.ts
@@ -6,5 +6,6 @@ export * from './chart-card'
 export * from './date-picker-range'
 export * from './kpi-tile'
 export * from './person-color'
+export * from './run-status-badge'
 export * from './severity-chip'
 export * from './ticket-status-badge'
