+++ web/components/ds/kpi-tile.tsx
@@ -13,7 +13,7 @@ export interface KpiTileProps extends React.ComponentProps<'div'> {
 
 export function KpiTile({ label, value, delta, deltaDirection = 'up', deltaTone = 'good', className, ...props }: KpiTileProps) {
   const Arrow = deltaDirection === 'up' ? ArrowUp : ArrowDown
-  const color = deltaTone === 'good' ? 'text-chart-2' : 'text-chart-4'
+  const color = deltaTone === 'good' ? 'text-chart-2' : 'text-destructive'
   return (
     <div className={cn('min-w-[158px] border p-3', className)} {...props}>
       <div className="text-[10px] font-medium uppercase tracking-[0.08em] text-muted-foreground">{label}</div>
